+++ API/LudumDare53.API/LudumDare53.API.csproj
@@ -8,10 +8,10 @@
 
     <ItemGroup>
       <PackageReference Include="MongoDB.Driver" Version="2.19.1" />
-      <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.5.0" />
       <PackageReference Include="Swashbuckle.AspNetCore.Swagger" Version="6.5.0" />
       <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.5.0" />
       <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.5.0" />
+      <PackageReference Include="Swashbuckle.Core" Version="5.6.0" />
     </ItemGroup>
 
 </Project>
