+++ API/LudumDare53.API/Models/Player.cs
@@ -1,6 +1,4 @@
-﻿using LudumDare53.API.Models.Request;
-
-namespace LudumDare53.API.Models;
+﻿namespace LudumDare53.API.Models;
 
 public class Player : Resource
 {
@@ -8,12 +6,4 @@ public class Player : Resource
     public int DaysCompleted { get; set; }
     public int TotalMoneyEarned { get; set; }
     public int DeliveriesMade { get; set; }
-    
-    public Player() { }
-
-    public Player(PlayerCreateRequest playerCreateRequest)
-    {
-        Id = playerCreateRequest.Id;
-        Name = playerCreateRequest.Name;
-    }
 }
