--- API/LudumDare53.API/Models/Request/PlayerCreateRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace LudumDare53.API.Models.Request;
-
-public class PlayerCreateRequest
-{
-    public string Id { get; set; }
-    public string Name { get; set; }
-}
