--- API/LudumDare53.API/Models/Request/ProcessGameResultsRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace LudumDare53.API.Models.Request;
-
-public class ProcessGameResultsRequest
-{
-    
-}
