+++ API/LudumDare53.API/Models/Resource.cs
@@ -8,9 +8,6 @@ public class Resource
     [BsonId]
     public string Id { get; set; }
 
-    [ConcurrencyCheck]
-    public int Version { get; set; }
-
     public DateTime CreatedAt { get; set; }
 
     public DateTime UpdatedAt { get; set; }
